requirements.txt
blinker==1.9.0
charset-normalizer==3.4.6
click==8.3.1
Flask==3.1.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
PyMySQL==1.1.2
python-dotenv==1.2.2
reportlab==4.4.10
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.7
